Add AppointmentTest.Refacory test. add new Test  HE003_TC-01,02,03.04,05,06,07
