+++ src/test/java/SingInTest.java
@@ -96,102 +96,4 @@ public class SingInTest extends BaseTest {
         assertEquals("Login failed! Please ensure the username and password are valid.", errorMessage);
     }
 
-    @Test
-    void test_HE_003_TC_1_Valid_Login_Make_appointment_Healt_Care_Program_Mediacare_Tokyo() {
-        var facilityTokyoCURAAssert = homePage
-                .makeAppointmentButton()
-                .enterUserName(TestValues.TEST_VALID_USERNAME)
-                .enterpassword(TestValues.TEST_VALID_PASSWORD)
-                .clickButtonOnLoginPage()
-                .clickTokyoCuraHealtCareCenter()
-                .clickApplyForHospitalreadmission()
-                .clickHealtCareProgramMedicare()
-                .clickVisitDate()
-                .nextMount()
-                .nextMount()
-                .nextMount()
-                .enterVisitDate()
-                .commentTextField(TestValues.TEST_COMMENT_TEXTFIELD)
-                .clickButtonAppointment()
-                .clickHamburgerMenuSummaryPage()
-                .historyButton()
-                .getTextFacilityTokyoCURA();
-        assertEquals(facilityTokyoCURAAssert,"Tokyo CURA Healthcare Center");
-
-    }
-    @Test
-    void test_HE_003_TC_2_Valid_Login_Make_appointment_Healt_Care_Program_Mediacare_Hongkong() {
-        var facilityHongkongCURAAssert = homePage
-                .makeAppointmentButton()
-                .enterUserName(TestValues.TEST_VALID_USERNAME)
-                .enterpassword(TestValues.TEST_VALID_PASSWORD)
-                .clickButtonOnLoginPage()
-                .clickHongkongCuraHealtCareCenter()
-                .clickApplyForHospitalreadmission()
-                .clickHealtCareProgramMedicare()
-                .clickVisitDate()
-                .nextMount()
-                .nextMount()
-                .nextMount()
-                .enterVisitDate()
-                .commentTextField(TestValues.TEST_COMMENT_TEXTFIELD)
-                .clickButtonAppointment()
-                .clickHamburgerMenuSummaryPage()
-                .historyButton()
-                .getTextfacilityHongkongCURA();
-
-        assertEquals(facilityHongkongCURAAssert,"Hongkong CURA Healthcare Center");
-
-    }
-
-    @Test
-    void test_HE_003_TC_3_Valid_Login_Make_appointment_Healt_Care_Program_Mediacare_Seoul() {
-        var facilitySeoulCURAAssert = homePage
-                .makeAppointmentButton()
-                .enterUserName(TestValues.TEST_VALID_USERNAME)
-                .enterpassword(TestValues.TEST_VALID_PASSWORD)
-                .clickButtonOnLoginPage()
-                .clickSeoulCuraHealtCareCenter()
-                .clickApplyForHospitalreadmission()
-                .clickHealtCareProgramMedicare()
-                .clickVisitDate()
-                .nextMount()
-                .nextMount()
-                .nextMount()
-                .enterVisitDate()
-                .commentTextField(TestValues.TEST_COMMENT_TEXTFIELD)
-                .clickButtonAppointment()
-                .clickHamburgerMenuSummaryPage()
-                .historyButton()
-                .getTEXTFacilitySeoulCURA();
-
-        assertEquals(facilitySeoulCURAAssert,"Seoul CURA Healthcare Center");
-
-    }
-    @Test
-    void test_HE_003_TC_4_No_Apply_for_hospital_readmission() {
-        var noApplyCheckBox = homePage
-                .makeAppointmentButton()
-                .enterUserName(TestValues.TEST_VALID_USERNAME)
-                .enterpassword(TestValues.TEST_VALID_PASSWORD)
-                .clickButtonOnLoginPage()
-                .clickSeoulCuraHealtCareCenter()
-                .clickHealtCareProgramMedicare()
-                .clickVisitDate()
-                .nextMount()
-                .nextMount()
-                .nextMount()
-                .enterVisitDate()
-                .commentTextField(TestValues.TEST_COMMENT_TEXTFIELD)
-                .clickButtonAppointment()
-                .clickHamburgerMenuSummaryPage()
-                .historyButton()
-                .getNoApplyCheckBox();
-
-        assertEquals(noApplyCheckBox,"No");
-
-    }
-
-
-
 }
